Building the cart feature
for this commit, ive been wokring to make the cart feature functional, such that when theuser selects add to cart, the product is reflected on the checout page. So far i am able to only display the alert, but when i try and display it on the page, it doesnt seem to work, so in my next commit, i will be working to resolve ths issue.